Make test reusable
